move categories into namespace admin
