+++ app/views/api/admin/categories/_category.json.jbuilder
