+++ app/views/api/admin/categories/categories.json.jbuilder
