+++ config/routes.rb
@@ -13,7 +13,9 @@ Rails.application.routes.draw do
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 
   namespace :api, defaults: { format: :json } do
-    resources :categories
+    namespace :admin do
+      resources :categories
+    end
     resources :companies,
               only: [:create, :update, :show]
   end
